data/requirements.txt
geopandas
seaborn
rasterio
numpy
matplotlib
shapely
scipy
h3
absl-py
rasterstats
duckdb
